backend/workers/mbox/requirements.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-firestore
google-cloud-pubsub
python-multipart
requests
